move test atom provider to testing folder
