+++ app/react/V2/Routes/Settings/Thesauri/specs/Thesauri.spec.tsx
@@ -17,8 +17,8 @@ import { createMemoryRouter, RouterProvider } from 'react-router-dom';
 import { Provider } from 'react-redux';
 import { has } from 'lodash';
 import { fromJS } from 'immutable';
-import { templatesAtom, TestAtomStoreProvider } from 'V2/atoms';
-import { LEGACY_createStore as reduxStoreCreator } from 'V2/testing';
+import { templatesAtom } from 'V2/atoms';
+import { LEGACY_createStore as reduxStoreCreator, TestAtomStoreProvider } from 'V2/testing';
 import { ThesauriList, thesauriLoader } from '../ThesauriList';
 import { EditThesaurus } from '../EditThesaurus';
 import { editThesaurusLoader } from '../helpers';
