+++ app/react/V2/testing/index.ts
@@ -1,2 +1,3 @@
 // eslint-disable-next-line camelcase
 export { LEGACY_createStore } from './testingHelpers';
+export { TestAtomStoreProvider } from './TestAtomStoreProvider';
